+++ AmplifyPlugins/API/Sources/AWSAPIPlugin/AWSAPIPlugin+Configure.swift
@@ -5,9 +5,8 @@
 // SPDX-License-Identifier: Apache-2.0
 //
 
-import Amplify
+@_spi(InternalAmplifyConfiguration) import Amplify
 import AWSPluginsCore
-import AppSyncRealTimeClient
 import AwsCommonRuntimeKit
 
 public extension AWSAPIPlugin {
@@ -20,8 +19,15 @@ public extension AWSAPIPlugin {
     /// - Throws:
     ///   - PluginError.pluginConfigurationError: If one of the required configuration values is invalid or empty
     func configure(using configuration: Any?) throws {
+        let dependencies: ConfigurationDependencies
+        if let configuration = configuration as? AmplifyOutputsData {
+            dependencies = try ConfigurationDependencies(configuration: configuration,
+                                                         apiAuthProviderFactory: authProviderFactory)
+        } else if let jsonValue = configuration as? JSONValue {
+            dependencies = try ConfigurationDependencies(configurationValues: jsonValue,
+                                                         apiAuthProviderFactory: authProviderFactory)
 
-        guard let jsonValue = configuration as? JSONValue else {
+        } else {
             throw PluginError.pluginConfigurationError(
                 "Could not cast incoming configuration to JSONValue",
                 """
@@ -33,8 +39,6 @@ public extension AWSAPIPlugin {
             )
         }
 
-        let dependencies = try ConfigurationDependencies(configurationValues: jsonValue,
-                                                         apiAuthProviderFactory: authProviderFactory)
         configure(using: dependencies)
 
         // Initialize SwiftSDK's CRT dependency for SigV4 signing functionality
@@ -53,18 +57,18 @@ extension AWSAPIPlugin {
     struct ConfigurationDependencies {
         let authService: AWSAuthServiceBehavior
         let pluginConfig: AWSAPICategoryPluginConfiguration
-        let subscriptionConnectionFactory: SubscriptionConnectionFactory
+        let appSyncRealTimeClientFactory: AppSyncRealTimeClientFactoryProtocol
         let logLevel: Amplify.LogLevel
 
         init(
             configurationValues: JSONValue,
             apiAuthProviderFactory: APIAuthProviderFactory,
             authService: AWSAuthServiceBehavior? = nil,
-            subscriptionConnectionFactory: SubscriptionConnectionFactory? = nil,
+            appSyncRealTimeClientFactory: AppSyncRealTimeClientFactoryProtocol? = nil,
             logLevel: Amplify.LogLevel? = nil
         ) throws {
             let authService = authService
-                ?? AWSAuthService()
+            ?? AWSAuthService()
 
             let pluginConfig = try AWSAPICategoryPluginConfiguration(
                 jsonValue: configurationValues,
@@ -72,15 +76,35 @@ extension AWSAPIPlugin {
                 authService: authService
             )
 
-            let subscriptionConnectionFactory = subscriptionConnectionFactory
-                ?? AWSSubscriptionConnectionFactory()
-
             let logLevel = logLevel ?? Amplify.Logging.logLevel
 
             self.init(
                 pluginConfig: pluginConfig,
                 authService: authService,
-                subscriptionConnectionFactory: subscriptionConnectionFactory,
+                appSyncRealTimeClientFactory: appSyncRealTimeClientFactory
+                ?? AppSyncRealTimeClientFactory(),
+                logLevel: logLevel
+            )
+        }
+
+        init(
+            configuration: AmplifyOutputsData,
+            apiAuthProviderFactory: APIAuthProviderFactory,
+            authService: AWSAuthServiceBehavior = AWSAuthService(),
+            appSyncRealTimeClientFactory: AppSyncRealTimeClientFactoryProtocol? = nil,
+            logLevel: Amplify.LogLevel = Amplify.Logging.logLevel
+        ) throws {
+            let pluginConfig = try AWSAPICategoryPluginConfiguration(
+                configuration: configuration,
+                apiAuthProviderFactory: apiAuthProviderFactory,
+                authService: authService
+            )
+
+            self.init(
+                pluginConfig: pluginConfig,
+                authService: authService,
+                appSyncRealTimeClientFactory: appSyncRealTimeClientFactory
+                ?? AppSyncRealTimeClientFactory(),
                 logLevel: logLevel
             )
         }
@@ -88,12 +112,12 @@ extension AWSAPIPlugin {
         init(
             pluginConfig: AWSAPICategoryPluginConfiguration,
             authService: AWSAuthServiceBehavior,
-            subscriptionConnectionFactory: SubscriptionConnectionFactory,
+            appSyncRealTimeClientFactory: AppSyncRealTimeClientFactoryProtocol,
             logLevel: Amplify.LogLevel
         ) {
             self.pluginConfig = pluginConfig
             self.authService = authService
-            self.subscriptionConnectionFactory = subscriptionConnectionFactory
+            self.appSyncRealTimeClientFactory = appSyncRealTimeClientFactory
             self.logLevel = logLevel
         }
 
@@ -108,8 +132,6 @@ extension AWSAPIPlugin {
     func configure(using dependencies: ConfigurationDependencies) {
         authService = dependencies.authService
         pluginConfig = dependencies.pluginConfig
-        subscriptionConnectionFactory = dependencies.subscriptionConnectionFactory
-        AppSyncRealTimeClient.logLevel = AppSyncRealTimeClient.LogLevel(
-            rawValue: dependencies.logLevel.rawValue) ?? .error
+        appSyncRealTimeClientFactory = dependencies.appSyncRealTimeClientFactory
     }
 }