+++ AmplifyPlugins/Auth/Sources/AWSCognitoAuthPlugin/AWSCognitoAuthPlugin+Configure.swift
@@ -6,7 +6,7 @@
 //
 
 import Foundation
-import Amplify
+@_spi(InternalAmplifyConfiguration) import Amplify
 
 import AWSCognitoIdentity
 import AWSCognitoIdentityProvider
@@ -24,16 +24,18 @@ extension AWSCognitoAuthPlugin {
     /// - Throws:
     ///   - PluginError.pluginConfigurationError: If one of the configuration values is invalid or empty
     public func configure(using configuration: Any?) throws {
-
-        guard let jsonValueConfiguration = configuration as? JSONValue else {
+        let authConfiguration: AuthConfiguration
+        if let configuration = configuration as? AmplifyOutputsData {
+            authConfiguration = try ConfigurationHelper.authConfiguration(configuration)
+            jsonConfiguration = ConfigurationHelper.createUserPoolJsonConfiguration(authConfiguration)
+        } else if let jsonValueConfiguration = configuration as? JSONValue {
+            jsonConfiguration = jsonValueConfiguration
+            authConfiguration = try ConfigurationHelper.authConfiguration(jsonValueConfiguration)
+        } else {
             throw PluginError.pluginConfigurationError(
                 AuthPluginErrorConstants.decodeConfigurationError.errorDescription,
                 AuthPluginErrorConstants.decodeConfigurationError.recoverySuggestion)
         }
-
-        jsonConfiguration = jsonValueConfiguration
-
-        let authConfiguration = try ConfigurationHelper.authConfiguration(jsonValueConfiguration)
 
         let credentialStoreResolver = CredentialStoreState.Resolver().eraseToAnyResolver()
         let credentialEnvironment = credentialStoreEnvironment(authConfiguration: authConfiguration)