+++ AmplifyPlugins/Auth/Tests/AuthHostedUIApp/AuthHostedUIApp/ContentView.swift
@@ -42,7 +42,6 @@ struct ContentView: View {
         }
         self.loading = false
     }
-
 }
 
 struct ContentView_Previews: PreviewProvider {